changed file locations, updated utils and Docker
