+++ aiInferenceModel/Dockerfile
@@ -1,5 +1,5 @@
 # Stage 1: Build Python environment with dependencies
-FROM python:3.8-slim AS builder
+FROM python:3.9-slim AS builder
 
 # Set the working directory to /app
 WORKDIR /app
@@ -10,9 +10,10 @@ COPY requirements.txt .
 # Install any needed packages specified in requirements.txt into a virtual environment
 RUN python -m venv /opt/venv && \
     /opt/venv/bin/pip install --no-cache-dir -r requirements.txt
+RUN apt-get clean && rm -rf /var/lib/apt/lists/* /tmp/* /var/tmp/* /var/cache/*
 
 # Stage 2: Create the final image
-FROM python:3.8-slim
+FROM python:3.9-slim
 
 # Set the working directory to /app
 WORKDIR /app
@@ -22,10 +23,6 @@ COPY --from=builder /opt/venv /opt/venv
 
 # Copy the application code
 COPY . .
-
-RUN apt-get clean && rm -rf /var/lib/apt/lists/* /tmp/* /var/tmp/* /var/cache/*
-
-RUN apt-get clean && rm -rf /var/lib/apt/lists/* /tmp/* /var/tmp/* /var/cache/*
 
 # Make port 80 available to the world outside this container
 EXPOSE 80
