+++ aiInferenceModel/ai_models/nuclickClassification.py
