+++ aiInferenceModel/ai_models/nuclickSegmentation.py
