+++ aiInferenceModel/ai_models/samMobile.py
@@ -1,17 +1,18 @@
-from segment_anything import SamPredictor, sam_model_registry
+from mobile_sam import SamPredictor
 import os
 import tempfile
 
 import cv2
 import numpy as np
 
-DEVICE = "cpu"  # Can be changed to a specific hardware device, e.g., "cuda:0" for GPU
-MODEL_TYPE = "vit_h"  # The type of model to use, e.g., "vit_h"
+DEVICE = "cuda:0"  # Can be changed to a specific hardware device, e.g., "cuda:0" for GPU
+MODEL_TYPE = "vit_t"  # The type of model to use, e.g., "vit_h"
 
 # Import necessary modules from the "segment_anything" package
 
 
-def run_ai_model_inferencing(json_data):
+def run_ai_model_inferencing(json_data, mobile_sam):
+
     """
     Run inference using a Semantic Segmentation model on input image data.
 
@@ -34,7 +35,7 @@ def run_ai_model_inferencing(json_data):
       Wan-Yen and Doll, Piotr and Girshick, Ross
     """
 
-    print('Running Segment anything facebook model')
+    print('Running MobileSAM model')
 
     # Extract image data and size information from JSON input
     image_data, size_data, foreground_data = json_data.get("image"), json_data.get("tilesize"), json_data.get("nuclei_location")
@@ -53,17 +54,8 @@ def run_ai_model_inferencing(json_data):
     image = cv2.imread(os.path.join(temp_dir, 'image.png'))
     print(f"image shape {image.shape}")
 
-    # Define the path to the pre-trained model checkpoint
-    CHECKPOINT_PATH = "/home/local/KHQ/s.erattakulangara/Documents/HistomicsTK_EKS/dsa-run-custom-ai-models/debug/weights/vit_h_lm.pth"
-
-    # Initialize the Semantic Segmentation model and move it to the specified
-    # device
-    sam = sam_model_registry[MODEL_TYPE](
-        checkpoint=CHECKPOINT_PATH).to(
-        device=DEVICE)
-
     # Generate masks for the input image using the model
-    predictor = SamPredictor(sam)
+    predictor = SamPredictor(mobile_sam)
     predictor.set_image(image)
     input_label = np.array([1])
 
@@ -95,8 +87,8 @@ def run_ai_model_inferencing(json_data):
 
 if __name__ == "__main__":
     # Example usage when run as a standalone script
-    image_file = "/home/local/KHQ/s.erattakulangara/Documents/HistomicsTK_EKS/dsa-run-custom-ai-models/debug/workspace/test_12.png"
-    image = cv2.imread(image_file)[:500,:500,:]
+    image_file = ""
+    image = cv2.imread(image_file)
     payload = {"image": image, "tilesize":(0,0,0,0,0,0),"nuclei_location":[[288, 302]]}
     output = run_ai_model_inferencing(payload)
     print(output)
